product_service/requirements.txt
flask
requests
redis
gunicorn
psycopg2-binary
mysql-connector-python